% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_global.R
\name{app_global}
\alias{app_global}
\title{Create any variables needed for the app}
\usage{
app_global()
}
\value{
creates variables and assigns them into the pickatable_global environment
}
\description{
Create any variables needed for the app
}
