% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{App server}
\usage{
app_server(input, output, session)
}
\description{
A function that takes arguments input, output and session. This function
is not designed to be called interactively but instead acts as the server
for a shiny app.
}
