% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_global.R
\name{unpack_global}
\alias{unpack_global}
\title{Unpack contents of pickatable global environment into another}
\usage{
unpack_global(into)
}
\arguments{
\item{into}{the environment to copy objects to}
}
\value{
invisibly copies the contents of the pickatable_global environment into another
}
\description{
Unpack contents of pickatable global environment into another
}
